float -> double 同じような計算をしてみたところ, float  time:37700 double time:8605 real   time:10953 doubleが圧倒的に早かった. doubleが早いとは思っていたが,ここまでとは… メモリなんかより,速度を重視するべきですね.

+++ source/manager/MouseMoveSight.d
@@ -28,10 +28,10 @@ class MouseMoveSight:UpdateAble
 			glfwSetCursorPos(window_,width / 2,height / 2);
 			mouselog_.reset();
 		}
-		float sideper = distance.x / width;
-		float  dipper = distance.y / height;
-		float sideangle = 60 * sideper * -1;///degree 反時計周り座標です
-		float  dipangle = 60 *  dipper * -1;
+		double sideper = distance.x / width;
+		double  dipper = distance.y / height;
+		double sideangle = 60 * sideper * -1;///degree 反時計周り座標です
+		double  dipangle = 60 *  dipper * -1;
 
 		debug
 		{
